requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Graph analysis
networkx>=3.1
python-louvain>=0.16  # Louvain community detection
leidenalg>=0.10.0     # Leiden community detection (improved Louvain)
python-igraph>=0.11.0 # Required for leidenalg

# Statistical methods
scikit-learn>=1.3.0  # Only for scripts/optimize_k.py (cross-validation)
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0

# Data formats
h5py>=3.9.0
tables>=3.8.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
mypy>=1.5.0
black>=23.7.0
ruff>=0.0.285

# Type hints
typing-extensions>=4.7.0

# Optional: Performance
# polars>=0.18.0
# dask[complete]>=2023.8.0
